move tsx files from assets folder + change index.ts to js
