+++ src/components/Dropdown/docs/index.story.tsx
@@ -21,7 +21,7 @@ import {
   optionsWithSections,
   optionsWithSubtitle,
 } from '../helpers';
-import { Heart } from '../../../assets/icons';
+import { Heart } from '../../../assets/icons/index';
 import * as Readme from '../../Dropdown/README.md';
 import { Placement } from 'wix-ui-core/dist/src/components/popover/Popover';
