+++ src/components/IconButton/IconButton.meta.tsx
@@ -1,7 +1,7 @@
 import { IconButton, Skins } from '.';
 import Registry from '@ui-autotools/registry';
 import * as React from 'react';
-import { ChevronLeft } from '../../assets/icons/ChevronLeft';
+import { ChevronLeft } from '../icons/ChevronLeft';
 
 const IconButtonMetadata = Registry.getComponentMetadata(IconButton);
 IconButtonMetadata.nonReactStrictModeCompliant = true;
