+++ src/components/Toast/CloseIconButton/CloseIconButton.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { IconButton } from '../../IconButton';
-import { Close as CloseIcon } from '../../../assets/icons';
+import { Close as CloseIcon } from '../../../assets/icons/index';
 import extendedStyles from './CloseIconButton.st.css';
 
 export const CloseIconButton = function(props) {
