+++ src/components/icons/DoubleChevronLeft.tsx
